+++ src/actor-sheet/actor-sheet.ts
@@ -1,26 +1,28 @@
-import type { CoreAlt } from "../actor/actor.interface";
-import { onManageActiveEffect } from "../shared/effects";
+import type { AttributeAlt, CoreAlt, Party } from "../actor/actor.interface";
 import { CheckDifficulty } from "../skill-check/skill-check.enum";
 import { SettingsService } from "../settings/settings.service";
-import { l } from "../shared/util";
+import { getGame, l } from "../shared/util";
+import { AugmentedData } from "./actor-sheet.interface";
+
+// Remove this when migrating to v10
+import tippy from "tippy.js";
 
 /**
  * Extend the basic ActorSheet with some very simple modifications
  * @extends {ActorSheet}
  */
-export class CryptomancerActorSheet extends ActorSheet {
+export class CryptomancerActorSheet extends ActorSheet<DocumentSheetOptions, AugmentedData> {
   private readonly settings = new SettingsService();
 
-  /** @override */
-  static get defaultOptions() {
+  static override get defaultOptions() {
     return mergeObject(super.defaultOptions, {
       classes: ["cryptomancer", "sheet", "actor"],
       template: "systems/cryptomancer/actor-sheet/actor-sheet.hbs",
-      width: 625,
-      height: 720,
+      width: 680,
+      height: 840,
       tabs: [
         {
-          navSelector: ".sheet-tabs",
+          navSelector: ".crypt-tabs",
           contentSelector: ".sheet-body",
           initial: "core",
         },
@@ -28,15 +30,13 @@ export class CryptomancerActorSheet extends ActorSheet {
     });
   }
 
-  /** @override */
-  get template() {
+  override get template() {
     return `systems/cryptomancer/actor-sheet/actor-sheet-${this.actor.data.type}.hbs`;
   }
 
   /* -------------------------------------------- */
 
-  /** @override */
-  async getData() {
+  override async getData() {
     // Retrieve the data structure from the base sheet. You can inspect or log
     // the context variable to see the structure, but some key properties for
     // sheets are the actor object, the data object, whether or not it's
@@ -46,135 +46,19 @@ export class CryptomancerActorSheet extends ActorSheet {
     return augmented;
   }
 
-  // TODO: I've moved the majority of this to the actor class, remove what I can
-  /**
-   * Add data that is specifically for rendering sheets, mainly
-   * inputs for partial components.
-   */
-  private augmentContext(
-    context: ActorSheet.Data<ActorSheet.Options>
-  ): ActorSheet.Data<ActorSheet.Options> {
-    // Prepare character data and items.
-    if (context.data.type === "character") {
-      this._prepareItems(context);
-      this.prepareCharacterData(context);
-      // Get configured check difficulty
-      context.data.data.checkDifficulty =
-        this.settings.getSetting("checkDifficulty") ??
-        CheckDifficulty.Challenging;
-    }
-
-    return context;
-  }
-
-  /**
-   * Organize and classify Items for Character sheets.
-   *
-   * @param {Object} actorData The actor to prepare.
-   *
-   * @return {undefined}
-   */
-  private prepareCharacterData(context: ActorSheet.Data<ActorSheet.Options>) {
-    if (context.data.type !== "character") return;
-
-    // Handle labels.
-    // Localize resources
-    context.data.data.healthPoints.label = l(
-      context.data.data.healthPoints.label
-    );
-    context.data.data.manaPoints.label = l(context.data.data.manaPoints.label);
-    context.data.data.upgradePoints.label = l(
-      context.data.data.upgradePoints.label
-    );
-
-    // Localize core, attribute, skill
-    for (let [coreKey, coreValue] of Object.entries(context.data.data.core)) {
-      coreValue.label = l(`Core.${coreKey}`);
-      for (let [attrKey, attrValue] of Object.entries(
-        (coreValue as CoreAlt).attributes
-      )) {
-        attrValue.label = l(`Attr.${attrKey}`);
-        if (attrValue.skills) {
-          for (let [skillKey, skillValue] of Object.entries(attrValue.skills)) {
-            skillValue.label = l(`Skill.${skillKey}`);
-          }
-        }
-      }
-    }
-  }
-
-  /**
-   * Organize and classify Items for Character sheets.
-   *
-   * @param {Object} actorData The actor to prepare.
-   *
-   * @return {undefined}
-   */
-  _prepareItems(context: ActorSheet.Data<ActorSheet.Options>) {
-    // // Initialize containers.
-    // const gear: ActorSheetItem[] = [];
-    // const talents: ActorSheetItem[] = [];
-    // // Iterate through items, allocating to containers
-    // for (let i of context.items) {
-    //   i.img = i.img || CONST.DEFAULT_TOKEN;
-    //   // Append to gear.
-    //   if (i.type === "item") {
-    //     gear.push(i);
-    //   }
-    //   // Append to features.
-    //   else if (i.type === "talent") {
-    //     talents.push(i);
-    //   }
-    // }
-  }
-
-  /* -------------------------------------------- */
-
-  /** @override */
-  activateListeners(html: JQuery<HTMLElement>) {
+  override activateListeners(html: JQuery<HTMLElement>) {
     super.activateListeners(html);
     this._activateCharacterListenters(html);
     this._activatePartyListenters(html);
-
-    // html.find(".talent-table .talent-row").each((_, row) => {
-    //   const id = row.dataset.talentId;
-    //   $(row)
-    //     .find(".action-button")
-    //     .on("click", (evt) => {
-    //       if (!id) {
-    //         return;
-    //       }
-    //       if (evt.target.classList.contains("view")) {
-    //         this.handleTalentAction(id, "view");
-    //       } else if (evt.target.classList.contains("edit")) {
-    //         this.handleTalentAction(id, "edit");
-    //       } else if (evt.target.classList.contains("delete")) {
-    //         this.handleTalentAction(id, "delete");
-    //       }
-    //     });
-    // });
+    tippy("[data-tooltip]", {
+      content: (reference) => {
+        return (reference as HTMLElement).dataset.tooltip as string;
+      },
+    });
 
     // -------------------------------------------------------------
     // Everything below here is only needed if the sheet is editable
     if (!this.isEditable) return;
-
-    // // Add Inventory Item
-    // html.find(".item-create").on("click", this._onItemCreate.bind(this));
-
-    // // Delete Inventory Item
-    // html.find(".item-delete").on("click", (ev) => {
-    //   const li = $(ev.currentTarget).parents(".item");
-    //   const item = this.actor.items.get(li.data("itemId"));
-    //   if (item) {
-    //     item.delete();
-    //     li.slideUp(200, () => this.render(false));
-    //   }
-    // });
-
-    // // Active Effect management
-    // html
-    //   .find(".effect-control")
-    //   .on("click", (ev) => onManageActiveEffect(ev, this.actor));
 
     // Drag events for macros.
     if (this.actor.isOwner) {
@@ -187,11 +71,86 @@ export class CryptomancerActorSheet extends ActorSheet {
     }
   }
 
+  /**
+   * Add data that is specifically for rendering sheets, mainly
+   * inputs for partial components.
+   */
+  private augmentContext(context: AugmentedData): AugmentedData {
+    this.prepareCharacterData(context);
+    return context;
+  }
+
+  /**
+   * Organize and classify Items for Character sheets.
+   *
+   * @param {Object} actorData The actor to prepare.
+   *
+   * @return {undefined}
+   */
+  private prepareCharacterData(context: AugmentedData) {
+    if (context.data.type !== "character") return;
+    // Get configured check difficulty
+    context.data.data.checkDifficulty = this.settings.getSetting("checkDifficulty") ?? CheckDifficulty.Challenging;
+
+    // Prep data for rendering
+    context.hpAttributeBar = {
+      color: "success",
+      max: context.data.data.healthPoints.max,
+      maxName: "data.healthPoints.max",
+      maxPlaceholder: "Max HP",
+      value: context.data.data.healthPoints.value,
+      valueName: "data.healthPoints.value",
+      valuePlaceholder: "HP",
+      class: "hp",
+      tooltip: "HP",
+    };
+    context.manaAttributeBar = {
+      color: "primary",
+      max: context.data.data.manaPoints.max,
+      maxName: "data.manaPoints.max",
+      maxPlaceholder: "Max Mana",
+      value: context.data.data.manaPoints.value,
+      valueName: "data.manaPoints.value",
+      valuePlaceholder: "MP",
+      class: "mp",
+      tooltip: "MP",
+    };
+
+    // Prep skills for rendering
+    context.skills = [];
+    Object.values(context.data.data.core).forEach((core) => {
+      Object.values(core.attributes).forEach((attr: AttributeAlt) => {
+        if (attr.skills) {
+          Object.values(attr.skills).forEach((skill) => {
+            skill.label = l(skill.label);
+            context.skills.push({
+              ...skill,
+              core: core.key,
+              attributeValue: attr.value,
+            });
+          });
+        }
+      });
+    });
+    context.skills.sort((a, b) => (a.label > b.label ? 1 : -1));
+
+    // Get all party sheets to select for this character
+    context.partyOptions = getGame().actors!.filter((a) => a.type === "party");
+    context.selectedParty = null;
+    const currentPartyId = context.data.data.biography.party;
+    if (currentPartyId) {
+      const selectedParty = context.partyOptions.find((p) => p.id === currentPartyId);
+      if (selectedParty) {
+        context.selectedParty = selectedParty.data.data as Party;
+      }
+    }
+  }
+
   private _activateCharacterListenters(html: JQuery<HTMLElement>): void {
     if (this.actor.data.type !== "character") return;
 
     // Rollable abilities.
-    html.find(".rollable").on("click", this._onRoll.bind(this));
+    html.find(".rollable").on("click", this.onRoll.bind(this));
 
     // Render the item sheet for viewing/editing prior to the editable check.
     html.find(".item-edit").on("click", (ev) => {
@@ -203,52 +162,60 @@ export class CryptomancerActorSheet extends ActorSheet {
     });
 
     // Add talent event listeners
-    html.find(".crypt-feature-list .action-button").on("click", (evt) => {
+    html.find(".crypt-item-table .action-button").on("click", (evt) => {
       const button = evt.target;
       const row = $(evt.currentTarget).parents(".item-row");
       const item = this.actor.items.get(row.data("itemId"));
       if (!item || !item.sheet) {
         return;
       }
-      if (
-        button.classList.contains("view") ||
-        button.classList.contains("edit")
-      ) {
+      if (button.classList.contains("view") || button.classList.contains("edit")) {
         item.sheet.render(true);
       } else if (button.classList.contains("delete")) {
         item.deleteDialog();
       }
     });
 
-    html
-      .find(".crypt-gear .trademark-items .trademark-item__action-button")
-      .on("click", (evt) => {
-        const button = evt.target;
-        const itemElement = $(evt.currentTarget).parents(".trademark-item");
-        const item = this.actor.items.get(itemElement.data("itemId"));
-        if (!item || !item.sheet) {
+    html.find(".crypt-item-table .avatar-wrapper").on("click", (evt) => {
+      const row = $(evt.currentTarget).parents(".item-row");
+      const item = this.actor.items.get(row.data("itemId"));
+      if (!item) {
+        return;
+      }
+      item.showChatMessage();
+    });
+
+    html.find(".crypt-item-table .name a").on("click", (evt) => {
+      const row = $(evt.currentTarget).parents(".item-row");
+      const item = this.actor.items.get(row.data("itemId"));
+      if (!item || !item.sheet) {
+        return;
+      }
+      item.sheet.render(true);
+    });
+
+    // Add party event listeners
+    html.find(".party-upgrade-points .party-action-button").on("click", async (evt) => {
+      const button = evt.target as HTMLButtonElement;
+      const partyId = button.dataset.partyId;
+      if (button.classList.contains("view") && partyId) {
+        const party = getGame().actors!.get(partyId);
+        if (!party || !party.sheet) {
           return;
         }
-        if (
-          button.classList.contains("view") ||
-          button.classList.contains("edit")
-        ) {
-          item.sheet.render(true);
-        } else if (button.classList.contains("delete")) {
-          item.deleteDialog();
-        }
-      });
+        party.sheet.render(true);
+      }
+    });
 
-    html
-      .find(".difficulty-selector")
-      .on("change", this._onDifficultySelect.bind(this));
+    // Difficulty selector
+    html.find(".difficulty-selector").on("change", this.onDifficultySelect.bind(this));
   }
 
   private _activatePartyListenters(html: JQuery<HTMLElement>): void {
     if (this.actor.data.type !== "party") return;
 
     // Operations skill checks
-    html.find(".rollable").on("click", this._onCellRoll.bind(this));
+    html.find(".rollable").on("click", this.onCellRoll.bind(this));
 
     // Cell time increment selector
     html.find(".time-increments__increment").on("click", (event) => {
@@ -264,44 +231,14 @@ export class CryptomancerActorSheet extends ActorSheet {
     });
   }
 
-  /**
-   * Handle creating a new Owned Item for the actor using initial data defined in the HTML dataset
-   * @param {Event} event   The originating click event
-   * @private
-   */
-  async _onItemCreate(event: JQuery.ClickEvent) {
-    event.preventDefault();
-    const header = event.currentTarget;
-    // Get the type of item to create.
-    const type = header.dataset.type;
-    // Grab any data associated with this control.
-    const data = duplicate(header.dataset);
-    // Initialize a default name.
-    const name = `New ${type.capitalize()}`;
-    // Prepare the item object.
-    const itemData = {
-      name: name,
-      type: type,
-      data: data,
-    };
-    // Remove the type from the dataset since it's in the itemData.type prop.
-    delete itemData.data["type"];
-
-    // Finally, create the item!
-    return await Item.create(itemData, { parent: this.actor });
-  }
-
-  _onDifficultySelect(event: JQuery.ChangeEvent) {
+  private onDifficultySelect(event: JQuery.ChangeEvent) {
     event.preventDefault();
     switch (event.target.id as string) {
       case "difficulty-trivial":
         this.settings.updateSetting("checkDifficulty", CheckDifficulty.Trivial);
         break;
       case "difficulty-challenging":
-        this.settings.updateSetting(
-          "checkDifficulty",
-          CheckDifficulty.Challenging
-        );
+        this.settings.updateSetting("checkDifficulty", CheckDifficulty.Challenging);
         break;
       case "difficulty-tough":
         this.settings.updateSetting("checkDifficulty", CheckDifficulty.Tough);
@@ -315,7 +252,7 @@ export class CryptomancerActorSheet extends ActorSheet {
    * @param {Event} event   The originating click event
    * @private
    */
-  _onRoll(event: JQuery.ClickEvent) {
+  private onRoll(event: JQuery.ClickEvent) {
     event.preventDefault();
     const element = event.currentTarget;
     const dataset = element.dataset;
@@ -328,12 +265,11 @@ export class CryptomancerActorSheet extends ActorSheet {
       rollCore,
       rollAttribute,
       rollSkill,
-      (this.settings.getSetting("checkDifficulty") as CheckDifficulty) ??
-        CheckDifficulty.Challenging
+      (this.settings.getSetting("checkDifficulty") as CheckDifficulty) ?? CheckDifficulty.Challenging
     );
   }
 
-  _onCellRoll(event: JQuery.ClickEvent) {
+  private onCellRoll(event: JQuery.ClickEvent) {
     if (this.actor.data.type !== "party") return;
     event.preventDefault();
     const element = event.currentTarget;
@@ -342,26 +278,5 @@ export class CryptomancerActorSheet extends ActorSheet {
     const index = dataset.index as "1" | "2" | "3";
 
     this.document.rollCellOperations(this.actor.data.data.cells[index]);
-  }
-
-  private handleTalentAction(
-    id: string,
-    action: "view" | "edit" | "delete"
-  ): void {
-    const talent = this.actor.items.get(id);
-    if (!talent) {
-      return;
-    }
-    switch (action) {
-      case "view":
-        talent.sheet?.render(true);
-        break;
-      case "edit":
-        talent.sheet?.render(true);
-        break;
-      case "delete":
-        talent.deleteDialog();
-        break;
-    }
   }
 }