+++ CHANGELOG.md
@@ -2,6 +2,55 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.5.0](https://github.com/OfficerHalf/cryptomancer-fvtt/compare/v0.4.0...v0.5.0) (2022-05-16)
+
+
+### Features
+
+* add equipment rules and spell info to item chat card ([23e7f25](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/23e7f25106f599c43ca3127bbc5d673b7ca64394))
+* basic support for chat cards from character sheet ([9defeaf](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/9defeaf770243f68c6152ffa231d9636be1966b2))
+
+
+### Bug Fixes
+
+* improve localization consistency, add document specific IDs to actor sheet inputs ([e34ecff](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/e34ecff0ba31f7eb2de3f6d5396e2d2b112f7240))
+
+## [0.4.0](https://github.com/OfficerHalf/cryptomancer-fvtt/compare/v0.3.0...v0.4.0) (2022-05-10)
+
+
+### Features
+
+* add equipment compendia ([6023649](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/6023649116619cf0ac1e38f64652c90dc2bb3668))
+* add equipment type and migrate trademarkItem items to equipment ([5f1e947](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/5f1e947e63299ccf955d5f8e327c8a64e947c359))
+* add/remove rules and qualities as array ([aad672c](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/aad672c6a2d004907b49da967bcf4d1ffc3d5d68))
+* redo item sheets to be more consistent with each other and use crypt-ui components ([06b00fe](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/06b00fed38e17768b8833416a8e3813e30e5b58f))
+* support for all item types and coin ([1b699a1](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/1b699a104f1fceb5116db5a70b9d2ed02b39f403))
+
+
+### Bug Fixes
+
+* add some extra space between other gear label and input ([b7ab6ff](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/b7ab6ff2c041eb6a019f7bca77533a8ea5b2cb6b))
+* add type="button" to buttons where appropriate ([e52c364](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/e52c3643500e38786b859dfeb90a68c4f7e1998c))
+* correct package.json version ([34bd033](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/34bd033614c2607cfa8cc71da5e89b2da3cec118))
+* select options have a legible text color ([f14848e](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/f14848eee2f1b1141f322493e13a031276bfc81d))
+
+## 0.3.0 (2022-05-04)
+
+### Features
+
+- cleanup templates, add linting setup ([7637457](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/7637457fe0489fa5858db27cae91248fa27f2dca))
+- first pass at crypt-ui and new Core character sheet ([29dd013](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/29dd013358b7b1ff8a483d53856a71667c3cfeb6))
+- formatting upgrade points section ([1750e00](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/1750e00aacafbb32613bb6c4ff6951c5dad9e9b4))
+- migrate to crypt-ui vars ([d495d62](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/d495d6249ad6059b2097a5d76ba39e243a4a8868))
+- use core-input and toggle-box from crypt-ui ([d49accd](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/d49accd7354630965f0ea4cf16a8f6a49e258648))
+- use named partials ([495fd3e](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/495fd3eb2276a7fa3d187168c154b0fb4aa7979a))
+- view actor sheet from character sheet ([2c8a4be](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/2c8a4be6d2899b56fc5b904456c4510320f2b809))
+
+### Bug Fixes
+
+- break and push work for endurance and willpower and toggles are on the sheet ([8f8d451](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/8f8d4516fe9f67d212ad2c6f9ef6a2e4dc7bad3b))
+- party sheet can be rendered again ([61dfa2e](https://github.com/OfficerHalf/cryptomancer-fvtt/commit/61dfa2ef733b8ddb6d3902887068890c5115d3ef))
+
 ## 0.2.0 (2022-04-09)
 
 ### Features